git-svn-id: https://localhost/svn/wxTTM/trunk@8259 7754d8ab-9b54-e011-a251-0001c006bdc7

+++ src/wxTTM/Dialogs/ReportListView.cpp
@@ -92,7 +92,7 @@ bool CReportListView::Edit(va_list vaList)
   pattern += "\\*.rep";
   
   _finddata_t findinfo; 
-  long h = _findfirst(pattern.data(), &findinfo);
+  intptr_t h = _findfirst(pattern.data(), &findinfo);
   if (h != -1L)
   {
     do
